server/requirements.txt
fastapi
uvicorn
jinja2
pandas
numpy
ujson
spotipy
google-genai